+++ code/source/draw.c
@@ -4,7 +4,7 @@
 
 #define OFFSET (h-1)/2
 
-void drawline(u8 x0, u8 y0, u8 x1, u8 y1, int color) 
+void drawline(int x0, int y0, int x1, int y1, int color) 
 {
 	if (x0 == x1)
 	{
@@ -61,7 +61,7 @@ void drawline(u8 x0, u8 y0, u8 x1, u8 y1, int color)
 	}
 }
 
-void drawRectXYXY(u8 x0, u8 y0, u8 x1, u8 y1, int color)
+void drawRectXYXY(int x0, int y0, int x1, int y1, int color)
 {
 	drawline(x1, y0, x0, y0, color);
 	drawline(x0, y1, x0, y0, color);
@@ -69,7 +69,7 @@ void drawRectXYXY(u8 x0, u8 y0, u8 x1, u8 y1, int color)
 	drawline(x1, y0, x1, y1, color);
 }
 
-void drawRectXYHW(u8 x, u8 y, u8 h, u8 w, int color)
+void drawRectXYHW(int x, int y, int h, int w, int color)
 {
 	drawline(x+h,     y,     x,   y, color);
 	drawline(  x, y+w-1,     x,   y, color);
@@ -77,7 +77,7 @@ void drawRectXYHW(u8 x, u8 y, u8 h, u8 w, int color)
 	drawline(x+h,     y,   x+h, y+w, color);
 }
 
-void drawCubeCentered(u8 x, u8 y, u8 h, int color)
+void drawCubeCentered(int x, int y, int h, int color)
 {
 	drawline(x-OFFSET, y+OFFSET, x+OFFSET, y+OFFSET, color);
 	drawline(x+OFFSET, y-OFFSET, x+OFFSET, y+OFFSET, color);
@@ -85,7 +85,7 @@ void drawCubeCentered(u8 x, u8 y, u8 h, int color)
 	drawline(x-OFFSET, y+OFFSET, x-OFFSET, y-OFFSET, color);
 }
 
-void drawRectXYHWfill(u8 x, u8 y, u8 h, u8 w, int color)
+void drawRectXYHWfill(int x, int y, int h, int w, int color)
 {
 	for (size_t i = 0; i <= h; i++) {
 		for (size_t j = 0; j <= w; j++){
