+++ code/source/paddle.h
@@ -6,10 +6,10 @@
 
 typedef struct
 {
-    u8 x;
-    u8 y;
-    u8 h; //height
-    u8 w; //width
+    int x;
+    int y;
+    int h; //height
+    int w; //width
 } Paddle;
 
 #endif //PONG_PADDLE_H
